requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.9.0
supabase==2.11.0
PyJWT==2.9.0
cryptography>=42.0.0
PyMuPDF==1.25.0
sentence-transformers==3.4.0
httpx==0.28.0
beautifulsoup4==4.12.0
